backend/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose[cryptography]
python-multipart
pydantic
